# go-server/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Syncra DMS backend API.",
    "title": "Syncra DMS API",
    "version": "0.1"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/api/auth/check-permission": {
      "post": {
        "description": "Trusted SvelteKit server endpoint. Requires an authenticated session.",
        "tags": [
          "me"
        ],
        "summary": "Check whether the current authenticated user has a permission.",
        "operationId": "checkPermission",
        "responses": {
          "200": {
            "description": "Permission decision was returned."
          },
          "400": {
            "description": "Invalid permission check request."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          }
        }
      }
    },
    "/api/auth/email-otp/send-verification-otp": {
      "post": {
        "description": "Trusted SvelteKit server action endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Send or rotate a six-digit email verification OTP.",
        "operationId": "sendVerificationOTP",
        "responses": {
          "200": {
            "description": "OTP was created when the account exists and is unverified."
          },
          "400": {
            "description": "Invalid OTP request."
          },
          "401": {
            "description": "Trusted internal request required."
          }
        }
      }
    },
    "/api/auth/email-otp/verify-email": {
      "post": {
        "description": "Trusted SvelteKit server action endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Confirm an email address with a six-digit OTP.",
        "operationId": "verifyEmailOTP",
        "responses": {
          "200": {
            "description": "Email was verified."
          },
          "400": {
            "description": "Invalid or expired verification code."
          },
          "401": {
            "description": "Trusted internal request required."
          }
        }
      }
    },
    "/api/auth/get-session": {
      "get": {
        "description": "Trusted SvelteKit server hook endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Load the current session from the auth.session_token cookie.",
        "operationId": "getSession",
        "responses": {
          "200": {
            "description": "Current authenticated session, or null when no valid session exists."
          },
          "401": {
            "description": "Trusted internal request required."
          }
        }
      }
    },
    "/api/auth/oauth/github/callback": {
      "post": {
        "description": "Trusted SvelteKit server endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Complete GitHub OAuth sign-in.",
        "operationId": "signInGitHubOAuth",
        "responses": {
          "200": {
            "description": "Authenticated session."
          },
          "401": {
            "description": "OAuth sign-in failed or trusted internal request required."
          }
        }
      }
    },
    "/api/auth/oauth/github/start": {
      "post": {
        "description": "Trusted SvelteKit server endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Start GitHub OAuth.",
        "operationId": "startGitHubOAuth",
        "responses": {
          "200": {
            "description": "GitHub authorization URL and state."
          },
          "401": {
            "description": "Trusted internal request required."
          },
          "503": {
            "description": "GitHub OAuth is not configured."
          }
        }
      }
    },
    "/api/auth/oauth/google/callback": {
      "post": {
        "description": "Trusted SvelteKit server endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Complete Google OAuth sign-in.",
        "operationId": "signInGoogleOAuth",
        "responses": {
          "200": {
            "description": "Authenticated session."
          },
          "401": {
            "description": "OAuth sign-in failed or trusted internal request required."
          }
        }
      }
    },
    "/api/auth/oauth/google/start": {
      "post": {
        "description": "Trusted SvelteKit server endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Start Google OAuth.",
        "operationId": "startGoogleOAuth",
        "responses": {
          "200": {
            "description": "Google authorization URL and state."
          },
          "401": {
            "description": "Trusted internal request required."
          },
          "503": {
            "description": "Google OAuth is not configured."
          }
        }
      }
    },
    "/api/auth/password-reset/confirm": {
      "post": {
        "description": "Trusted SvelteKit server action endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Confirm password reset with emailed token.",
        "operationId": "confirmPasswordReset",
        "responses": {
          "200": {
            "description": "Password was reset and existing sessions were revoked."
          },
          "400": {
            "description": "Invalid or expired password reset token."
          },
          "401": {
            "description": "Trusted internal request required."
          }
        }
      }
    },
    "/api/auth/password-reset/request": {
      "post": {
        "description": "Trusted SvelteKit server action endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Request a password reset email.",
        "operationId": "requestPasswordReset",
        "responses": {
          "200": {
            "description": "Reset request accepted."
          },
          "400": {
            "description": "Invalid password reset request."
          },
          "401": {
            "description": "Trusted internal request required."
          }
        }
      }
    },
    "/api/auth/sign-in/email": {
      "post": {
        "description": "Trusted SvelteKit server action endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Sign in with email and password.",
        "operationId": "signInEmail",
        "responses": {
          "200": {
            "description": "Authenticated session."
          },
          "401": {
            "description": "Invalid email or password, or trusted internal request required."
          },
          "403": {
            "description": "Email is not verified."
          }
        }
      }
    },
    "/api/auth/sign-out": {
      "post": {
        "description": "Trusted SvelteKit server action endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Sign out the current session.",
        "operationId": "signOut",
        "responses": {
          "200": {
            "description": "Session was deleted when present."
          },
          "401": {
            "description": "Trusted internal request required."
          }
        }
      }
    },
    "/api/auth/sign-up/email": {
      "post": {
        "description": "Trusted SvelteKit server action endpoint.",
        "tags": [
          "auth"
        ],
        "summary": "Sign up with email and password.",
        "operationId": "signUpEmail",
        "responses": {
          "200": {
            "description": "User account was created or already exists and email verification is required."
          },
          "400": {
            "description": "Invalid sign-up request."
          },
          "401": {
            "description": "Trusted internal request required."
          }
        }
      }
    },
    "/api/document-folders": {
      "post": {
        "description": "Trusted SvelteKit server endpoint. Requires document.create scoped to the organization unit.",
        "tags": [
          "documentFolders"
        ],
        "summary": "Create a root or child document folder.",
        "operationId": "createDocumentFolder",
        "responses": {
          "201": {
            "description": "Document folder created."
          },
          "400": {
            "description": "Invalid document folder request."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "document.create permission required."
          },
          "404": {
            "description": "Organization unit or parent folder not found."
          },
          "409": {
            "description": "Active document folder name already exists or parent belongs to another organization unit."
          }
        }
      }
    },
    "/api/document-folders/tree": {
      "get": {
        "description": "Trusted SvelteKit server endpoint. Requires document.view scoped to the organization unit.",
        "tags": [
          "documentFolders"
        ],
        "summary": "List active document folders for an organization unit as a hierarchy.",
        "operationId": "listDocumentFolderTree",
        "parameters": [
          {
            "type": "string",
            "name": "organizationUnitId",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Active document folder hierarchy."
          },
          "400": {
            "description": "Invalid organization unit id."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "document.view permission required."
          },
          "404": {
            "description": "Organization unit not found."
          }
        }
      }
    },
    "/api/document-folders/{id}": {
      "patch": {
        "description": "Trusted SvelteKit server endpoint. Requires document.update scoped to the folder's organization unit.",
        "tags": [
          "documentFolders"
        ],
        "summary": "Update an active document folder's details.",
        "operationId": "updateDocumentFolder",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Document folder updated."
          },
          "400": {
            "description": "Invalid document folder request."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "404": {
            "description": "Document folder not found or inaccessible."
          },
          "409": {
            "description": "Active document folder name already exists or organization unit change was requested."
          }
        }
      }
    },
    "/api/document-folders/{id}/archive": {
      "post": {
        "description": "Trusted SvelteKit server endpoint. Requires document.delete scoped to the folder's organization unit.",
        "tags": [
          "documentFolders"
        ],
        "summary": "Archive an active document folder subtree and active documents under it.",
        "operationId": "archiveDocumentFolder",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Document folder subtree archived."
          },
          "400": {
            "description": "Invalid document folder id."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "404": {
            "description": "Document folder not found or inaccessible."
          }
        }
      }
    },
    "/api/document-folders/{id}/contents": {
      "get": {
        "description": "Trusted SvelteKit server endpoint. Requires document.view scoped to the folder's organization unit. Document metadata excludes storage keys.",
        "tags": [
          "documentFolders"
        ],
        "summary": "List active child folders and documents in a folder.",
        "operationId": "listDocumentFolderContents",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Active child folders and document metadata were listed."
          },
          "400": {
            "description": "Invalid document folder id."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "404": {
            "description": "Document folder not found or inaccessible."
          }
        }
      }
    },
    "/api/document-folders/{id}/parent": {
      "patch": {
        "description": "Trusted SvelteKit server endpoint. Requires document.update scoped to the folder's organization unit.",
        "tags": [
          "documentFolders"
        ],
        "summary": "Move an active document folder to another parent or root.",
        "operationId": "moveDocumentFolder",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Document folder moved."
          },
          "400": {
            "description": "Invalid move request."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "404": {
            "description": "Document folder not found or inaccessible, or parent not found."
          },
          "409": {
            "description": "Move would create a cycle, cross units, or duplicate an active folder name."
          }
        }
      }
    },
    "/api/documents/upload": {
      "post": {
        "description": "Trusted SvelteKit server endpoint. Requires document.create scoped to the folder's organization unit. Document metadata excludes storage keys.",
        "consumes": [
          "multipart/form-data"
        ],
        "tags": [
          "documents"
        ],
        "summary": "Upload a document file into an active document folder.",
        "operationId": "uploadDocument",
        "parameters": [
          {
            "type": "string",
            "name": "folderId",
            "in": "formData",
            "required": true
          },
          {
            "type": "file",
            "name": "file",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Document metadata was created and the file was stored."
          },
          "400": {
            "description": "Invalid multipart upload, document folder id, or missing file."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "404": {
            "description": "Document folder not found or inaccessible."
          },
          "409": {
            "description": "Active document with the same file hash already exists in the folder."
          },
          "413": {
            "description": "Uploaded document exceeds the configured maximum size."
          },
          "415": {
            "description": "Uploaded document MIME type is not allowed."
          }
        }
      }
    },
    "/api/documents/{id}": {
      "get": {
        "description": "Task 9 placeholder. Trusted SvelteKit server endpoint.",
        "tags": [
          "documents"
        ],
        "summary": "Get document metadata.",
        "operationId": "getDocument",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "401": {
            "description": "Trusted internal request required."
          },
          "501": {
            "description": "Document metadata endpoint is not implemented until Task 9."
          }
        }
      },
      "patch": {
        "description": "Task 9 placeholder. Trusted SvelteKit server endpoint.",
        "tags": [
          "documents"
        ],
        "summary": "Update document metadata.",
        "operationId": "updateDocument",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "401": {
            "description": "Trusted internal request required."
          },
          "501": {
            "description": "Document update endpoint is not implemented until Task 9."
          }
        }
      }
    },
    "/api/documents/{id}/archive": {
      "post": {
        "description": "Task 9 placeholder. Trusted SvelteKit server endpoint.",
        "tags": [
          "documents"
        ],
        "summary": "Archive an active document.",
        "operationId": "archiveDocument",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "401": {
            "description": "Trusted internal request required."
          },
          "501": {
            "description": "Document archive endpoint is not implemented until Task 9."
          }
        }
      }
    },
    "/api/documents/{id}/download": {
      "get": {
        "description": "Task 9 placeholder. Trusted SvelteKit server endpoint.",
        "tags": [
          "documents"
        ],
        "summary": "Download a stored document file.",
        "operationId": "downloadDocument",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "401": {
            "description": "Trusted internal request required."
          },
          "501": {
            "description": "Document download endpoint is not implemented until Task 9."
          }
        }
      }
    },
    "/api/groups": {
      "get": {
        "description": "Trusted SvelteKit server endpoint. Requires group.view.",
        "tags": [
          "groups"
        ],
        "summary": "List groups.",
        "operationId": "listGroups",
        "responses": {
          "200": {
            "description": "Groups were listed."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "group.view permission required."
          }
        }
      },
      "post": {
        "description": "Trusted SvelteKit server endpoint. Requires group.create.",
        "tags": [
          "groups"
        ],
        "summary": "Create a group.",
        "operationId": "createGroup",
        "responses": {
          "201": {
            "description": "Group was created."
          },
          "400": {
            "description": "Invalid group request."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "group.create permission required."
          },
          "409": {
            "description": "Group code already exists."
          }
        }
      }
    },
    "/api/groups/{id}": {
      "get": {
        "description": "Trusted SvelteKit server endpoint. Requires group.view.",
        "tags": [
          "groups"
        ],
        "summary": "Get a group.",
        "operationId": "getGroup",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Group was loaded."
          },
          "400": {
            "description": "Invalid group id."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "group.view permission required."
          },
          "404": {
            "description": "Group not found."
          }
        }
      },
      "delete": {
        "description": "Trusted SvelteKit server endpoint. Requires group.delete.",
        "tags": [
          "groups"
        ],
        "summary": "Delete an unused group.",
        "operationId": "deleteGroup",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Group was deleted."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "group.delete permission required."
          },
          "404": {
            "description": "Group not found."
          },
          "409": {
            "description": "Group has users or roles."
          }
        }
      },
      "patch": {
        "description": "Trusted SvelteKit server endpoint. Requires group.update.",
        "tags": [
          "groups"
        ],
        "summary": "Update a group.",
        "operationId": "updateGroup",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Group was updated."
          },
          "400": {
            "description": "Invalid group request."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "group.update permission required."
          },
          "404": {
            "description": "Group not found."
          },
          "409": {
            "description": "Group code already exists."
          }
        }
      }
    },
    "/api/groups/{id}/roles": {
      "post": {
        "description": "Trusted SvelteKit server endpoint. Requires group.assign_roles.",
        "tags": [
          "groups"
        ],
        "summary": "Assign a scoped role to a group.",
        "operationId": "assignGroupRole",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Group role assignment was created."
          },
          "400": {
            "description": "Invalid group role request."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "group.assign_roles permission required."
          },
          "404": {
            "description": "Group, role, or organization unit not found."
          }
        }
      }
    },
    "/api/groups/{id}/roles/{assignmentId}": {
      "delete": {
        "description": "Trusted SvelteKit server endpoint. Requires group.assign_roles.",
        "tags": [
          "groups"
        ],
        "summary": "Remove a group role assignment.",
        "operationId": "removeGroupRole",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "assignmentId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Group role assignment was removed."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "group.assign_roles permission required."
          },
          "404": {
            "description": "Group role assignment not found."
          }
        }
      }
    },
    "/api/groups/{id}/users": {
      "post": {
        "description": "Trusted SvelteKit server endpoint. Requires group.manage_users.",
        "tags": [
          "groups"
        ],
        "summary": "Add a user to a group.",
        "operationId": "addGroupUser",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Group membership was created."
          },
          "400": {
            "description": "Invalid group user request."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "group.manage_users permission required."
          },
          "404": {
            "description": "Group or user not found."
          }
        }
      }
    },
    "/api/groups/{id}/users/{userId}": {
      "delete": {
        "description": "Trusted SvelteKit server endpoint. Requires group.manage_users.",
        "tags": [
          "groups"
        ],
        "summary": "Remove a user from a group.",
        "operationId": "removeGroupUser",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Group membership was removed."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          },
          "403": {
            "description": "group.manage_users permission required."
          },
          "404": {
            "description": "Group membership not found."
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "description": "Trusted SvelteKit server endpoint. Requires an authenticated session.",
        "tags": [
          "me"
        ],
        "summary": "Get the current authenticated user.",
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "Current user was loaded."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          }
        }
      }
    },
    "/api/me/permissions": {
      "get": {
        "description": "Trusted SvelteKit server endpoint. Requires an authenticated session.",
        "tags": [
          "me"
        ],
        "summary": "List effective permissions for the current authenticated user.",
        "operationId": "getMyPermissions",
        "responses": {
          "200": {
            "description": "Effective permissions were listed."
          },
          "401": {
            "description": "Authenticated session or trusted internal request required."
          }
        }
      }
    },
...
  }
}